# Conflicts:
#	src/main/java/com/aldogg/sorter/long_/st/RadixByteSorterLong.java
#	src/main/java/com/aldogg/sorter/shared/Section.java

+++ src/test/java/com/aldogg/sorter/generators/FloatGenerator.java
@@ -31,7 +31,7 @@ public class FloatGenerator {
         } else {
             int range = (int) (params.limitHigh - params.limitLow);
             for (int i = 0; i < size; ++i) {
-                int randomInt = params.random.nextInt(range) + params.limitLow;
+                int randomInt = params.random.nextInt(range) + (int) params.limitLow;
                 v[i] = randomInt;
             }
         }
