+++ src/test/java/com/aldogg/sorter/generators/GeneratorParams.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class GeneratorParams {
     public int size;
-    public int limitLow;
+    public long limitLow;
     public long limitHigh;
     public Random random;
     public GeneratorFunctions function;
